split into postgres & shared process event test
